refactor(persistent): forcing blockdata value
Changing the way of the blockdata value if forced (when changed by a
debug stick) from metadata to persistent data container. Now each
location of changed block is registered in the chunk data container.
This change has been done like that the value of a blockdata will
continue to be forced after the restart of the server.

In addition to that, two commands has been added to control that :
- `/pds chunk info`: to see all block that the values of their blockdata
  will be forced in the chunk
- `pds chunk clear`: to clear ALL block that the value of their
  blockdata has to be forced in the chunk