refactor(debustick): update data storage
The data storage has changed to prevent in the future some errors when
deleting properties (or rename it) while a debug stick has it as current
property.
Cause of serialization, it was not possible to only exclude the current
property, so if the current property was deleted, the debug stick was
not usable anymore.

Now the data storage is more flexible to get data. It continues to use
the bytes array but in a different way :
- The first element of the byte array is always (whatever the type of
  debug stick) the ordinal of the current property minus 128.
  e.g. If the current property is "ORIENTABLE" the ordinal is 0. 0 - 128
  = -128.
  The minus 128 is here to be sure that there is no byte overflow if the
  number of property go to more than 128. Like this there 256 max
  properties.
- The fourth last elements of the byte array is only if the debug stick
  is a basic debug stick. They represent the durability of the basic
  debug stick converted to a byte array (an int to a byte array of size
  4).
So if the size of the bytes arrays is :
= 1 -> infinite debug stick
= 5 -> basic debug stick