feat(lang): better ways to change language
To do that, all the lang files has been moved into a new folder called
'lang'. And the 'messages.yml' has been renamed to 'message_en.yml' for
more logical.

Now the plugin need to be reloaded (via '/pds reload-config') for the
messages config to be reload to. This has been done for more
optimisation (the plugin will not parse the yml file each time a message
is send).