perf: change way to cache the blocks to not update
The olds ways to cache the blocks to not update was to use the Guava
cache, but these was not enough performants for the use.

Now, when a chunk is loaded the plugin will cache all the blocks from
the cache to not update.
When the chunk is unloaded, it will removes the blocks of the chunk
from the cache.

Fix: #29