fix(bisected): cause exception on top door
When editing the property bisected on the top part of the door a
ClassCastException was thrown.
This exception was thrown because the changment of the property on this
part of the door, makes the door disapear and when the plugin was trying
to get the new value of the property on the block by casting it to
bisected, the block casted was an air block.

Fixed this by disabling the physic when applying the new value of the
property, like that the door will not break.
It also fix a duplication bug that appears when the door was broken, it
appears 2 items of the door.