feat(property): add waxed property
This property is applied to sign and if set to true, the sign will not
be editable.

To do this, the interface `PropertyAction` has been update to overload
the method `getValue` to use the block as parameter instead of the block
data.
This has been done because the block can provide more data than the
block data.
By default, the block data method is used when implenting the interface,
because it's a easier to use.
But for the sign, we use the block state that cannot be provided by the
block data.