fix(property): phase property was bad registered
When adding the property phase to the list of usables properties, the
blockdata class used to check if a block coulb has this property edited
was "SculkShrieker" and not "SculkSensor".
This has consequence to the schulk shrieker block has the possibility to
have this property edited and so when casting it to "Sculk Sensor" in
the property action a ClassCastException was thrown.

Fix: #35